requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-mail
mysql-connector-python
python-dotenv
werkzeug
alembic
cryptography
bcrypt==4.0.1
PyJWT
passlib[bcrypt]==1.7.4
razorpay
python-barcode[images]
opencv-python
pyzbar
Pillow
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
boto3